+++ linux-sample/sys_cb_linux.c
@@ -1,4 +1,4 @@
-#include "kii_thing_if_environment_impl.h"
+#include "sys_cb_impl.h"
 
 #include <kii_core_secure_socket.h>
 #include <kii_socket_impl.h>
